FIR: JavaSymbolProviderWrapper -> JavaSymbolProvider
